added validation for Komitent and Roba

+++ WPFPresentation/ViewModels/Komitent/IzmeniKomitentaViewModel.cs
@@ -1,14 +1,17 @@
 ﻿using Application.DTOs;
+using FluentValidation;
 using System.Collections.ObjectModel;
 using System.Windows.Input;
 using WPFPresentation.Commands;
 using WPFPresentation.Services;
+using WPFPresentation.Validators;
 
 namespace WPFPresentation.ViewModels.Komitent
 {
     public class IzmeniKomitentaViewModel : BaseViewModel
     {
         private KomitentService _komitentService;
+        private IValidator<KomitentDTO> _validator;
 
         private string? _searchString;
         public string? SearchString
@@ -52,6 +55,19 @@ namespace WPFPresentation.ViewModels.Komitent
             }
         }
 
+        private string? _validation;
+        public string? Validation
+        {
+            get
+            {
+                return _validation;
+            }
+            set
+            {
+                _validation = value;
+                OnPropertyChanged(nameof(Validation));
+            }
+        }
 
         public ICommand FindKomitentCommand { get; set; }
         public ICommand UpdateKomitentCommand { get; set; }
@@ -62,6 +78,7 @@ namespace WPFPresentation.ViewModels.Komitent
             _komitentService = new KomitentService();
             SelectedKomitent = new KomitentDTO();
             Komitenti = new ObservableCollection<KomitentDTO>();
+            _validator = new KomitentValidator();
             FindKomitentCommand = new RelayCommand(async (obj) => await FindKomitents(obj));
             UpdateKomitentCommand = new RelayCommand(async (obj) => await UpdateKomitent(obj));
             DeleteKomitentCommand = new RelayCommand(async (obj) => await DeleteKomitent(obj));
@@ -69,6 +86,11 @@ namespace WPFPresentation.ViewModels.Komitent
 
         private async Task DeleteKomitent(object obj)
         {
+            if (SelectedKomitent!.Id == 0)
+            {
+                Validation = "Morate izabrati komitenta";
+                return;
+            }
             await _komitentService.DeleteKomitent(SelectedKomitent!);
             SelectedKomitent = new KomitentDTO();
             await FindKomitents(obj);
@@ -76,13 +98,31 @@ namespace WPFPresentation.ViewModels.Komitent
 
         private async Task UpdateKomitent(object obj)
         {
+            if (SelectedKomitent!.Id == 0)
+            {
+                Validation = "Morate izabrati komitenta";
+                return;
+            }
+            var result = _validator.Validate(SelectedKomitent!);
+            if (!result.IsValid)
+            {
+                Validation = string.Join("\n", result.Errors.Select(error => error.ErrorMessage));
+                return;
+            }
             await _komitentService.UpdateKomitent(SelectedKomitent!);
             SelectedKomitent = new KomitentDTO();
         }
 
         private async Task FindKomitents(object obj)
         {
-            var komitenti = await _komitentService.FindKomitents(SearchString!);
+            IEnumerable<KomitentDTO> komitenti = new List<KomitentDTO>();
+            if (string.IsNullOrWhiteSpace(SearchString))
+            {
+                komitenti = await _komitentService.GetKomitents();
+                Komitenti = new ObservableCollection<KomitentDTO>(komitenti);
+                return;
+            }
+            komitenti = await _komitentService.FindKomitents(SearchString);
             Komitenti = new ObservableCollection<KomitentDTO>(komitenti);
         }
     }
